move the time-tag edit mode into state namespace.
